#1 removed unnecessary files
